+++ Fossil/src/Core.cpp
@@ -2,18 +2,23 @@
 
 #include "Settings.h"
 #include "Log.h"
+#include "Game.h"
 #include "../test/TestHeader.h"
 
-//	Instantiate global logger
+//	Instantiate global objects
 Log logger;
+Game game;
 
 int main() {
 	//	Log message noting that the game started
 	logger.logMessage(". . . Starting . . . \n", "gamelogs.txt", true);
 
-	//	Initialize the settings vectors
-	initializeSettings();
+	//	Initialize the game
+	game.InitializeGame();
 
 	//	Log message noting that the game has exited
 	logger.logMessage("\n . . . Ending . . .", "gamelogs.txt", false);
+
+	//	For testing to allow console to stay open
+	std::cin.get();
 }