+++ Fossil/src/Fossil.h
@@ -6,3 +6,4 @@
 #include <vector>
 #include "Settings.h"
 #include "Log.h"
+#include "Settings.h"