+++ Fossil/src/Game.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "Settings.h"
+#include "Log.h"
+
+class Game
+{
+public:
+	Game();
+	virtual ~Game();
+
+	void InitializeGame();
+	void Run();
+};
+
+extern Game game;