+++ Fossil/src/Log.h
@@ -3,6 +3,7 @@
 #include <iostream>
 #include <string>
 #include <fstream>
+#include <Windows.h>
 
 class Log
 {
@@ -12,6 +13,15 @@ public:
 
 	//	Function to log a message to a text file
 	void logMessage(std::string message, std::string filePath, bool clear);
+
+	//	function to log message to console
+	void logInfo(std::string message);
+
+	//	Function to log warning to console
+	void logWarn(std::string message);
+
+	//	Function to log warning to console
+	void logError(std::string message);
 };
 
 extern Log logger;