+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Changed FileLogger to use strings for printing instead of const char*
 - Creation of FileReader class
 	- Implementation of getSettings function to retrieve the settings of the game stored in gameprefs file
+- Creation of multiplayerprefs file to store multiplayer settings
+	- I am not near the implementation of multiplayer yet, but I figured it'd be good to get my idea concreted and to get general architecture in place.
 
 **Next:**
 - Open window using display settings specified in gameprefs.txt